requirements.txt
Pillow==8.3.2
PyYAML==5.4.1
aiohttp==3.7.4.post0
aioreloader==0.4.0
uvloop
